third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGCircleElement-dom-r-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'r' attribute of the SVGCircleElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS circleElement.getAttribute('r') is "1"
PASS circleElement.getAttribute('r') is "150"
PASS successfullyParsed is true

TEST COMPLETE
